made EmailSize into it's own component from Settings.tsx
