+++ src/components/Settings/Settings.tsx
@@ -1,20 +1,13 @@
-import { useEffect } from 'react'
-import Switch from '@mui/material/Switch'
 import Modal from '@mui/material/Modal'
 import FormGroup from '@mui/material/FormGroup'
-import FormControlLabel from '@mui/material/FormControlLabel'
-import Box from '@mui/material/Box'
-import MenuItem from '@mui/material/MenuItem'
 import {
   selectIsSettingsOpen,
   setIsSettingsOpen,
-  setEmailFetchSize,
-  selectEmailListSize,
 } from '../../Store/utilsSlice'
 import { useAppDispatch, useAppSelector } from '../../Store/hooks'
 import * as S from './SettingsStyles'
-import * as global from '../../constants/globalConstants'
 import ShowAvatar from './SettingsOptions/ShowAvatar/ShowAvatar'
+import EmailSize from './SettingsOptions/EmailSize/EmailSize'
 
 const handleClose = (dispatch: Function) => dispatch(setIsSettingsOpen(false))
 
@@ -23,12 +16,6 @@ const SETTINGS = 'Settings'
 const Settings = () => {
   const dispatch = useAppDispatch()
   const isSettingsOpen = useAppSelector(selectIsSettingsOpen)
-  const fetchCount = useAppSelector(selectEmailListSize)
-
-  const handleEmailListSizeChange = (e: any) => {
-    localStorage.setItem('fetchSize', e.target.value)
-    dispatch(setEmailFetchSize(e.target.value))
-  }
 
   return (
     <Modal
@@ -42,28 +29,7 @@ const Settings = () => {
         <S.SettingsContainer>
           <FormGroup>
             <ShowAvatar />
-
-            <FormControlLabel
-              label="Emails Fetched at a time"
-              control={
-                <Box sx={{ minWidth: 25, marginRight: 0.5 }}>
-                  <S.StyledSelect
-                    value={fetchCount}
-                    inputProps={{
-                      name: 'emailSize',
-                      id: 'uncontrolled-native',
-                    }}
-                    onChange={handleEmailListSizeChange}
-                  >
-                    {global.POSSIBLE_FETCH_SIZES.map((size) => (
-                      <MenuItem key={size} value={size}>
-                        {size}
-                      </MenuItem>
-                    ))}
-                  </S.StyledSelect>
-                </Box>
-              }
-            />
+            <EmailSize />
           </FormGroup>
         </S.SettingsContainer>
       </S.Dialog>
